opusdec: Silence clang 10 conversion warning
src/opusdec.c:581:28: warning: implicit conversion from 'int'
    to 'float' changes value from 2147483647 to 2147483648
    [-Wimplicit-int-float-conversion]

+++ src/opusdec.c
@@ -578,7 +578,7 @@ static int decode_cb(decode_cb_ctx *ctx, OpusMSDecoder *decoder, void *pcm,
    (void)nchannels;
    (void)li;
    lost = ctx->loss_percent>0
-    && 100*((float)rand())/RAND_MAX<ctx->loss_percent;
+    && 100*(float)rand()/(float)RAND_MAX<ctx->loss_percent;
    switch (format)
    {
       case OP_DEC_FORMAT_SHORT:
